+++ docs/changelog.rst
@@ -10,6 +10,7 @@ v2.1.9 General maintenance
 - add support for getting team colors from incomplete or partial team
   names or from team names containing typos or extra words
 - make fuzzy string matching for event names case-insensitive
+- set missing lap start times to pit out time where possible (fixes #29)
 
 
 v2.1.8 Add temporary support for sprint qualifying